requirements.txt
customtkinter
pillow
numpy
